Add files via upload
Main and SubMain are soley for testing.
User is unchanged.
Administrator and Doctor have to be filled with your methods. Both now extends form Staff, which extends from User.
Sample code on how to retrieve info from staffList, medicineList and patientList can be found in the Main and SubMain.

Download both folders and dump them into the same project.